long double + cam

+++ ShaderMandelbrot/ShaderMandelbrot/Vec2.cpp
@@ -1,6 +1,6 @@
 #include "Vec2.h"
 
-Vec2::Vec2(double x1, double y1) {
+Vec2::Vec2(long double x1, long double y1) {
 	x = x1;
 	y = y1;
 }
@@ -10,10 +10,10 @@ Vec2 operator+(const Vec2& a, const Vec2& b) {
 Vec2 operator-(const Vec2& a, const Vec2& b) {
 	return Vec2(a.x - b.x, a.y - b.y);
 }
-Vec2 operator*(const Vec2& a, double k) {
+Vec2 operator*(const Vec2& a, long double k) {
 	return Vec2(a.x*k, a.y*k);
 }
-Vec2 operator/(const Vec2& a, double k) {
+Vec2 operator/(const Vec2& a, long double k) {
 	return Vec2(a.x / k, a.y / k);
 }
 Vec2 operator+=(Vec2& a, const Vec2& b) {
