+++ ShaderMandelbrot/ShaderMandelbrot/Vec2.h
@@ -1,14 +1,14 @@
 #pragma once
 class Vec2 {
 public:
-	double x=0, y=0;
+	long double x=0, y=0;
 	Vec2() {};
-	Vec2(double x1, double y1);
+	Vec2(long double x1, long double y1);
 };
 
 Vec2 operator+(const Vec2& a, const Vec2& b);
 Vec2 operator-(const Vec2& a, const Vec2& b);
-Vec2 operator*(const Vec2& a, double k);
-Vec2 operator/(const Vec2& a, double k);
+Vec2 operator*(const Vec2& a, long double k);
+Vec2 operator/(const Vec2& a, long double k);
 Vec2 operator+=(Vec2& a, const Vec2& b);
 Vec2 operator-=(Vec2& a, const Vec2& b);
